Move to using docker image for lambda
