+++ docker-compose.override.yml
@@ -16,4 +16,4 @@ services:
       - ./lambda_functions/v2/functions/documents/app:/var/task/app
   localstack:
     volumes:
-      - ./lambda_functions/v2/functions/csv-sync/main.zip:/tmp/localstack/main.zip
+      - ./lambda_functions/v2/functions/csv-forwarder/main.zip:/tmp/localstack/main.zip
