--- lambda_functions/v2/functions/csv-sync/go.mod
@@ -1,10 +1,0 @@
-module github.com/ministryofjustice/opg-data-deputy-reporting/lambda_functions/v2/functions/csv-sync
-
-go 1.18
-
-require (
-	github.com/aws/aws-lambda-go v1.32.0
-	github.com/aws/aws-sdk-go v1.44.31
-)
-
-require github.com/jmespath/go-jmespath v0.4.0 // indirect
